Add dllexport keyword for API class, remove some warning, Add DYE preprocessor
warning 4251, warning probably loss of data (size_t -> int)

+++ src/IComponent.cpp
@@ -361,7 +361,7 @@ namespace DYE
 
 	void ReusablePool::ResizePool(std::size_t _size)
 	{
-		int diff = _size - m_MaxSize;
+		long long diff = _size - m_MaxSize;
 		if (diff > 0)
 		{
 			expandPool(diff);
